portfolio_page/requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
Flask-WTF==1.2.1
Flask-Security-Too==5.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0 